+++ src/langchain_google_alloydb_pg/vectorstore.py
@@ -48,6 +48,7 @@ class AlloyDBVectorStore(VectorStore):
         engine: AlloyDBEngine,
         embedding_service: Embeddings,
         table_name: str,
+        schema_name: str = "public",
         content_column: str = "content",
         embedding_column: str = "embedding",
         metadata_columns: List[str] = [],
@@ -65,10 +66,10 @@ class AlloyDBVectorStore(VectorStore):
             engine (AlloyDBEngine): Connection pool engine for managing connections to AlloyDB database.
             embedding_service (Embeddings): Text embedding model to use.
             table_name (str): Name of the existing table or the table to be created.
+            schema_name (str, optional): Name of the database schema. Defaults to "public".
             content_column (str): Column that represent a Document’s page_content. Defaults to "content".
             embedding_column (str): Column for embedding vectors. The embedding is generated from the document value. Defaults to "embedding".
             metadata_columns (List[str]): Column(s) that represent a document's metadata.
-            ignore_metadata_columns (List[str]): Column(s) to ignore in pre-existing tables for a document's metadata. Can not be used with metadata_columns. Defaults to None.
             id_column (str): Column that represents the Document's id. Defaults to "langchain_id".
             metadata_json_column (str): Column to store metadata as JSON. Defaults to "langchain_metadata".
             distance_strategy (DistanceStrategy): Distance strategy to use for vector similarity search. Defaults to COSINE_DISTANCE.
@@ -89,6 +90,7 @@ class AlloyDBVectorStore(VectorStore):
         self.engine = engine
         self.embedding_service = embedding_service
         self.table_name = table_name
+        self.schema_name = schema_name
         self.content_column = content_column
         self.embedding_column = embedding_column
         self.metadata_columns = metadata_columns
@@ -106,6 +108,7 @@ class AlloyDBVectorStore(VectorStore):
         engine: AlloyDBEngine,
         embedding_service: Embeddings,
         table_name: str,
+        schema_name: str = "public",
         content_column: str = "content",
         embedding_column: str = "embedding",
         metadata_columns: List[str] = [],
@@ -124,6 +127,7 @@ class AlloyDBVectorStore(VectorStore):
             engine (AlloyDBEngine): Connection pool engine for managing connections to AlloyDB database.
             embedding_service (Embeddings): Text embedding model to use.
             table_name (str): Name of an existing table.
+            schema_name (str, optional): Name of the database schema. Defaults to "public".
             content_column (str): Column that represent a Document’s page_content. Defaults to "content".
             embedding_column (str): Column for embedding vectors. The embedding is generated from the document value. Defaults to "embedding".
             metadata_columns (List[str]): Column(s) that represent a document's metadata.
@@ -144,7 +148,7 @@ class AlloyDBVectorStore(VectorStore):
                 "Can not use both metadata_columns and ignore_metadata_columns."
             )
         # Get field type information
-        stmt = f"SELECT column_name, data_type FROM information_schema.columns WHERE table_name = '{table_name}'"
+        stmt = f"SELECT column_name, data_type FROM information_schema.columns WHERE table_name = '{table_name}' AND table_schema = '{schema_name}'"
         results = await engine._afetch(stmt)
         columns = {}
         for field in results:
@@ -192,6 +196,7 @@ class AlloyDBVectorStore(VectorStore):
             engine,
             embedding_service,
             table_name,
+            schema_name,
             content_column,
             embedding_column,
             metadata_columns,
@@ -210,6 +215,7 @@ class AlloyDBVectorStore(VectorStore):
         engine: AlloyDBEngine,
         embedding_service: Embeddings,
         table_name: str,
+        schema_name: str = "public",
         content_column: str = "content",
         embedding_column: str = "embedding",
         metadata_columns: List[str] = [],
@@ -229,17 +235,18 @@ class AlloyDBVectorStore(VectorStore):
             engine (AlloyDBEngine): Connection pool engine for managing connections to AlloyDB database.
             embedding_service (Embeddings): Text embedding model to use.
             table_name (str): Name of an existing table.
-            content_column (str): Column that represent a Document’s page_content. Defaults to "content".
-            embedding_column (str): Column for embedding vectors. The embedding is generated from the document value. Defaults to "embedding".
-            metadata_columns (List[str]): Column(s) that represent a document's metadata.
-            ignore_metadata_columns (List[str]): Column(s) to ignore in pre-existing tables for a document's metadata. Can not be used with metadata_columns. Defaults to None.
-            id_column (str): Column that represents the Document's id. Defaults to "langchain_id".
-            metadata_json_column (str): Column to store metadata as JSON. Defaults to "langchain_metadata".
-            distance_strategy (DistanceStrategy): Distance strategy to use for vector similarity search. Defaults to COSINE_DISTANCE.
-            k (int): Number of Documents to return from search. Defaults to 4.
-            fetch_k (int): Number of Documents to fetch to pass to MMR algorithm.
-            lambda_mult (float): Number between 0 and 1 that determines the degree of diversity among the results with 0 corresponding to maximum diversity and 1 to minimum diversity. Defaults to 0.5.
-            index_query_options (QueryOptions): Index query option.
+            schema_name (str, optional): Name of the database schema. Defaults to "public".
+            content_column (str, optional): Column that represent a Document’s page_content. Defaults to "content".
+            embedding_column (str, optional): Column for embedding vectors. The embedding is generated from the document value. Defaults to "embedding".
+            metadata_columns (List[str]): Column(s) that represent a document's metadata. Defaults to an empty list.
+            ignore_metadata_columns (Optional[List[str]]): Column(s) to ignore in pre-existing tables for a document's metadata. Can not be used with metadata_columns. Defaults to None.
+            id_column (str, optional): Column that represents the Document's id. Defaults to "langchain_id".
+            metadata_json_column (str, optional): Column to store metadata as JSON. Defaults to "langchain_metadata".
+            distance_strategy (DistanceStrategy, optional): Distance strategy to use for vector similarity search. Defaults to COSINE_DISTANCE.
+            k (int, optional): Number of Documents to return from search. Defaults to 4.
+            fetch_k (int, optional): Number of Documents to fetch to pass to MMR algorithm. Defaults to 20.
+            lambda_mult (float, optional): Number between 0 and 1 that determines the degree of diversity among the results with 0 corresponding to maximum diversity and 1 to minimum diversity. Defaults to 0.5.
+            index_query_options (Optional[QueryOptions], optional): Index query option. Defaults to None.
 
         Returns:
             AlloyDBVectorStore
@@ -248,6 +255,7 @@ class AlloyDBVectorStore(VectorStore):
             engine,
             embedding_service,
             table_name,
+            schema_name,
             content_column,
             embedding_column,
             metadata_columns,
@@ -286,7 +294,7 @@ class AlloyDBVectorStore(VectorStore):
                 if len(self.metadata_columns) > 0
                 else ""
             )
-            insert_stmt = f'INSERT INTO "{self.table_name}"({self.id_column}, {self.content_column}, {self.embedding_column}{metadata_col_names}'
+            insert_stmt = f'INSERT INTO "{self.schema_name}"."{self.table_name}"({self.id_column}, {self.content_column}, {self.embedding_column}{metadata_col_names}'
             values = {"id": id, "content": content, "embedding": str(embedding)}
             values_stmt = "VALUES (:id, :content, :embedding"
 
@@ -372,7 +380,7 @@ class AlloyDBVectorStore(VectorStore):
             return False
 
         id_list = ", ".join([f"'{id}'" for id in ids])
-        query = f'DELETE FROM "{self.table_name}" WHERE {self.id_column} in ({id_list})'
+        query = f'DELETE FROM "{self.schema_name}"."{self.table_name}" WHERE {self.id_column} in ({id_list})'
         await self.engine._aexecute(query)
         return True
 
@@ -391,6 +399,7 @@ class AlloyDBVectorStore(VectorStore):
         embedding: Embeddings,
         engine: AlloyDBEngine,
         table_name: str,
+        schema_name: str = "public",
         metadatas: Optional[List[dict]] = None,
         ids: Optional[List[str]] = None,
         content_column: str = "content",
@@ -408,14 +417,15 @@ class AlloyDBVectorStore(VectorStore):
             embedding (Embeddings): Text embedding model to use.
             engine (AlloyDBEngine): Connection pool engine for managing connections to AlloyDB database.
             table_name (str): Name of an existing table.
-            metadatas (Optional[List[dict]]): List of metadatas to add to table records.
-            ids: (Optional[List[str]]): List of IDs to add to table records.
-            content_column (str): Column that represent a Document’s page_content. Defaults to "content".
-            embedding_column (str): Column for embedding vectors. The embedding is generated from the document value. Defaults to "embedding".
-            metadata_columns (List[str]): Column(s) that represent a document's metadata.
-            ignore_metadata_columns (List[str]): Column(s) to ignore in pre-existing tables for a document's metadata. Can not be used with metadata_columns. Defaults to None.
-            id_column (str): Column that represents the Document's id. Defaults to "langchain_id".
-            metadata_json_column (str): Column to store metadata as JSON. Defaults to "langchain_metadata".
+            schema_name (str, optional): Name of the database schema. Defaults to "public".
+            metadatas (Optional[List[dict]], optional): List of metadatas to add to table records. Defaults to None.
+            ids: (Optional[List[str]], optional): List of IDs to add to table records. Defaults to None.
+            content_column (str, optional): Column that represent a Document’s page_content. Defaults to "content".
+            embedding_column (str, optional): Column for embedding vectors. The embedding is generated from the document value. Defaults to "embedding".
+            metadata_columns (List[str], optional): Column(s) that represent a document's metadata. Defaults to an empty list.
+            ignore_metadata_columns (Optional[List[str]], optional): Column(s) to ignore in pre-existing tables for a document's metadata. Can not be used with metadata_columns. Defaults to None.
+            id_column (str, optional): Column that represents the Document's id. Defaults to "langchain_id".
+            metadata_json_column (str, optional): Column to store metadata as JSON. Defaults to "langchain_metadata".
 
         Returns:
             AlloyDBVectorStore
@@ -424,6 +434,7 @@ class AlloyDBVectorStore(VectorStore):
             engine,
             embedding,
             table_name,
+            schema_name,
             content_column,
             embedding_column,
             metadata_columns,
@@ -441,6 +452,7 @@ class AlloyDBVectorStore(VectorStore):
         embedding: Embeddings,
         engine: AlloyDBEngine,
         table_name: str,
+        schema_name: str = "public",
         ids: Optional[List[str]] = None,
         content_column: str = "content",
         embedding_column: str = "embedding",
@@ -457,14 +469,14 @@ class AlloyDBVectorStore(VectorStore):
             embedding (Embeddings): Text embedding model to use.
             engine (AlloyDBEngine): Connection pool engine for managing connections to AlloyDB database.
             table_name (str): Name of an existing table.
-            metadatas (Optional[List[dict]]): List of metadatas to add to table records.
-            ids: (Optional[List[str]]): List of IDs to add to table records.
-            content_column (str): Column that represent a Document’s page_content. Defaults to "content".
-            embedding_column (str): Column for embedding vectors. The embedding is generated from the document value. Defaults to "embedding".
-            metadata_columns (List[str]): Column(s) that represent a document's metadata.
-            ignore_metadata_columns (List[str]): Column(s) to ignore in pre-existing tables for a document's metadata. Can not be used with metadata_columns. Defaults to None.
-            id_column (str): Column that represents the Document's id. Defaults to "langchain_id".
-            metadata_json_column (str): Column to store metadata as JSON. Defaults to "langchain_metadata".
+            schema_name (str, optional): Name of the database schema. Defaults to "public".
+            ids: (Optional[List[str]], optional): List of IDs to add to table records. Defaults to None.
+            content_column (str, optional): Column that represent a Document’s page_content. Defaults to "content".
+            embedding_column (str, optional): Column for embedding vectors. The embedding is generated from the document value. Defaults to "embedding".
+            metadata_columns (List[str], optional): Column(s) that represent a document's metadata. Defaults to an empty list.
+            ignore_metadata_columns (Optional[List[str]], optional): Column(s) to ignore in pre-existing tables for a document's metadata. Can not be used with metadata_columns. Defaults to None.
+            id_column (str, optional): Column that represents the Document's id. Defaults to "langchain_id".
+            metadata_json_column (str, optional): Column to store metadata as JSON. Defaults to "langchain_metadata".
 
         Returns:
             AlloyDBVectorStore
@@ -474,6 +486,7 @@ class AlloyDBVectorStore(VectorStore):
             engine,
             embedding,
             table_name,
+            schema_name,
             content_column,
             embedding_column,
             metadata_columns,
@@ -493,6 +506,7 @@ class AlloyDBVectorStore(VectorStore):
         embedding: Embeddings,
         engine: AlloyDBEngine,
         table_name: str,
+        schema_name: str = "public",
         metadatas: Optional[List[dict]] = None,
         ids: Optional[List[str]] = None,
         content_column: str = "content",
@@ -510,14 +524,15 @@ class AlloyDBVectorStore(VectorStore):
             embedding (Embeddings): Text embedding model to use.
             engine (AlloyDBEngine): Connection pool engine for managing connections to AlloyDB database.
             table_name (str): Name of an existing table.
-            metadatas (Optional[List[dict]]): List of metadatas to add to table records.
-            ids: (Optional[List[str]]): List of IDs to add to table records.
-            content_column (str): Column that represent a Document’s page_content. Defaults to "content".
-            embedding_column (str): Column for embedding vectors. The embedding is generated from the document value. Defaults to "embedding".
-            metadata_columns (List[str]): Column(s) that represent a document's metadata.
-            ignore_metadata_columns (List[str]): Column(s) to ignore in pre-existing tables for a document's metadata. Can not be used with metadata_columns. Defaults to None.
-            id_column (str): Column that represents the Document's id. Defaults to "langchain_id".
-            metadata_json_column (str): Column to store metadata as JSON. Defaults to "langchain_metadata".
+            schema_name (str, optional): Name of the database schema. Defaults to "public".
+            metadatas (Optional[List[dict]], optional): List of metadatas to add to table records. Defaults to None.
+            ids: (Optional[List[str]], optional): List of IDs to add to table records. Defaults to None.
+            content_column (str, optional): Column that represent a Document’s page_content. Defaults to "content".
+            embedding_column (str, optional): Column for embedding vectors. The embedding is generated from the document value. Defaults to "embedding".
+            metadata_columns (List[str], optional): Column(s) that represent a document's metadata. Defaults to empty list.
+            ignore_metadata_columns (Optional[List[str]], optional): Column(s) to ignore in pre-existing tables for a document's metadata. Can not be used with metadata_columns. Defaults to None.
+            id_column (str, optional): Column that represents the Document's id. Defaults to "langchain_id".
+            metadata_json_column (str, optional): Column to store metadata as JSON. Defaults to "langchain_metadata".
 
         Returns:
             AlloyDBVectorStore
@@ -527,6 +542,7 @@ class AlloyDBVectorStore(VectorStore):
             embedding,
             engine,
             table_name,
+            schema_name,
             metadatas=metadatas,
             content_column=content_column,
             embedding_column=embedding_column,
@@ -546,6 +562,7 @@ class AlloyDBVectorStore(VectorStore):
         embedding: Embeddings,
         engine: AlloyDBEngine,
         table_name: str,
+        schema_name: str = "public",
         ids: Optional[List[str]] = None,
         content_column: str = "content",
         embedding_column: str = "embedding",
@@ -562,14 +579,14 @@ class AlloyDBVectorStore(VectorStore):
             embedding (Embeddings): Text embedding model to use.
             engine (AlloyDBEngine): Connection pool engine for managing connections to AlloyDB database.
             table_name (str): Name of an existing table.
-            metadatas (Optional[List[dict]]): List of metadatas to add to table records.
-            ids: (Optional[List[str]]): List of IDs to add to table records.
-            content_column (str): Column that represent a Document’s page_content. Defaults to "content".
-            embedding_column (str): Column for embedding vectors. The embedding is generated from the document value. Defaults to "embedding".
-            metadata_columns (List[str]): Column(s) that represent a document's metadata.
-            ignore_metadata_columns (List[str]): Column(s) to ignore in pre-existing tables for a document's metadata. Can not be used with metadata_columns. Defaults to None.
-            id_column (str): Column that represents the Document's id. Defaults to "langchain_id".
-            metadata_json_column (str): Column to store metadata as JSON. Defaults to "langchain_metadata".
+            schema_name (str, optional): Name of the database schema. Defaults to "public".
+            ids: (Optional[List[str]], optional): List of IDs to add to table records. Defaults to None.
+            content_column (str, optional): Column that represent a Document’s page_content. Defaults to "content".
+            embedding_column (str, optional): Column for embedding vectors. The embedding is generated from the document value. Defaults to "embedding".
+            metadata_columns (List[str], optional): Column(s) that represent a document's metadata. Defaults to an empty list.
+            ignore_metadata_columns (Optional[List[str]], optional): Column(s) to ignore in pre-existing tables for a document's metadata. Can not be used with metadata_columns. Defaults to None.
+            id_column (str, optional): Column that represents the Document's id. Defaults to "langchain_id".
+            metadata_json_column (str, optional): Column to store metadata as JSON. Defaults to "langchain_metadata".
 
         Returns:
             AlloyDBVectorStore
@@ -579,6 +596,7 @@ class AlloyDBVectorStore(VectorStore):
             embedding,
             engine,
             table_name,
+            schema_name,
             content_column=content_column,
             embedding_column=embedding_column,
             metadata_columns=metadata_columns,
@@ -603,7 +621,7 @@ class AlloyDBVectorStore(VectorStore):
         search_function = self.distance_strategy.search_function
 
         filter = f"WHERE {filter}" if filter else ""
-        stmt = f"SELECT *, {search_function}({self.embedding_column}, '{embedding}') as distance FROM \"{self.table_name}\" {filter} ORDER BY {self.embedding_column} {operator} '{embedding}' LIMIT {k};"
+        stmt = f"SELECT *, {search_function}({self.embedding_column}, '{embedding}') as distance FROM \"{self.schema_name}\".\"{self.table_name}\" {filter} ORDER BY {self.embedding_column} {operator} '{embedding}' LIMIT {k};"
         if self.index_query_options:
             query_options_stmt = f"SET LOCAL {self.index_query_options.to_string()};"
             results = await self.engine._afetch_with_query_options(
@@ -918,9 +936,9 @@ class AlloyDBVectorStore(VectorStore):
             await self.adrop_vector_index()
             return
 
-        # Create `postgres_ann` extension when a `ScaNN` index is applied
+        # Create `alloydb_scann` extension when a `ScaNN` index is applied
         if isinstance(index, ScaNNIndex):
-            await self.engine._aexecute("CREATE EXTENSION IF NOT EXISTS postgres_ann")
+            await self.engine._aexecute("CREATE EXTENSION IF NOT EXISTS alloydb_scann")
             function = index.distance_strategy.scann_index_function
         else:
             function = index.distance_strategy.index_function
@@ -931,7 +949,7 @@ class AlloyDBVectorStore(VectorStore):
             if index.name == None:
                 index.name = self.table_name + DEFAULT_INDEX_NAME_SUFFIX
             name = index.name
-        stmt = f"CREATE INDEX {'CONCURRENTLY' if concurrently else ''} {name} ON \"{self.table_name}\" USING {index.index_type} ({self.embedding_column} {function}) {params} {filter};"
+        stmt = f"CREATE INDEX {'CONCURRENTLY' if concurrently else ''} {name} ON \"{self.schema_name}\".\"{self.table_name}\" USING {index.index_type} ({self.embedding_column} {function}) {params} {filter};"
         if concurrently:
             await self.engine._aexecute_outside_tx(stmt)
         else:
@@ -961,7 +979,7 @@ class AlloyDBVectorStore(VectorStore):
         query = f"""
         SELECT tablename, indexname
         FROM pg_indexes
-        WHERE tablename = '{self.table_name}' AND indexname = '{index_name}';
+        WHERE tablename = '{self.table_name}' AND schemaname = '{self.schema_name}' AND indexname = '{index_name}';
         """
         results = await self.engine._afetch(query)
         return bool(len(results) == 1)