+++ CHANGELOG.md
@@ -1,5 +1,37 @@
 # Changelog
 
+## [0.6.0](https://github.com/googleapis/langchain-google-alloydb-pg-python/compare/v0.5.0...v0.6.0) (2024-09-03)
+
+
+### ⚠ BREAKING CHANGES
+
+* Update `postgres_ann` to `alloydb_scann` ([#212](https://github.com/googleapis/langchain-google-alloydb-pg-python/issues/212))
+
+### Features
+
+* Add ScaNN index query options ([#198](https://github.com/googleapis/langchain-google-alloydb-pg-python/issues/198)) ([0813867](https://github.com/googleapis/langchain-google-alloydb-pg-python/commit/081386779ef80e419cb734d21673a8feb195a271))
+
+
+### Bug Fixes
+
+* Fix `QueryOptions` not applied to similarity search bug ([#203](https://github.com/googleapis/langchain-google-alloydb-pg-python/issues/203)) ([df206d0](https://github.com/googleapis/langchain-google-alloydb-pg-python/commit/df206d090d3b1cc83770d0052095362002ff377c))
+* Update `postgres_ann` to `alloydb_scann` ([#212](https://github.com/googleapis/langchain-google-alloydb-pg-python/issues/212)) ([783dcad](https://github.com/googleapis/langchain-google-alloydb-pg-python/commit/783dcad96b3cf0ae2ac11af826b71dff67a47f1f))
+
+
+### Documentation
+
+* Add index choosing guide ([#200](https://github.com/googleapis/langchain-google-alloydb-pg-python/issues/200)) ([dd91d54](https://github.com/googleapis/langchain-google-alloydb-pg-python/commit/dd91d54f0bfc2942704537b062862b038f9f3891))
+* Added info on how to modify existing libraries ([#205](https://github.com/googleapis/langchain-google-alloydb-pg-python/issues/205)) ([202b086](https://github.com/googleapis/langchain-google-alloydb-pg-python/commit/202b086678bcd55ccc2e69592d8db7fbfe4f5acb))
+* Added vector store initialization from documents ([#196](https://github.com/googleapis/langchain-google-alloydb-pg-python/issues/196)) ([823821c](https://github.com/googleapis/langchain-google-alloydb-pg-python/commit/823821c647b2e0fd6a7a44d4a2ed6d11ffb1037a))
+* Create functions need pre-existing tables ([#202](https://github.com/googleapis/langchain-google-alloydb-pg-python/issues/202)) ([a2e6c0a](https://github.com/googleapis/langchain-google-alloydb-pg-python/commit/a2e6c0a1d1ef2887d3fc9edac7cd62a0864f5a17))
+* Rename index guide ([#201](https://github.com/googleapis/langchain-google-alloydb-pg-python/issues/201)) ([2436c5b](https://github.com/googleapis/langchain-google-alloydb-pg-python/commit/2436c5b2120d69938644e20181559aa62b95fc84))
+* Update index tuning samples to include advanced indexes ([#197](https://github.com/googleapis/langchain-google-alloydb-pg-python/issues/197)) ([da52bc4](https://github.com/googleapis/langchain-google-alloydb-pg-python/commit/da52bc421da1ca8cad78184566426b22941b09af))
+
+
+### Miscellaneous Chores
+
+* Release 0.6.0 ([#217](https://github.com/googleapis/langchain-google-alloydb-pg-python/issues/217)) ([3d96251](https://github.com/googleapis/langchain-google-alloydb-pg-python/commit/3d9625131e3412e6d2958856224ef251baa625bd))
+
 ## [0.5.0](https://github.com/googleapis/langchain-google-alloydb-pg-python/compare/v0.4.1...v0.5.0) (2024-07-23)